chore: split lib into per-concern subdirectories
Reorganize flat lib/*.ts into hierarchical lib/<topic>/ folders,
one file per concern with a co-located test.

- worktree -> cleanup/list/remove/guards/uncommitted
- branch   -> cleanup/list/remove/guards
- merged   -> index (composite) + ancestor/squash/rebase/untouched
- flags    -> parse/help
- ui       -> color/logo/relpath/status
- agent    -> session/scope
- git      -> exec
- resolve-base, testing also moved into their own folders

logo.ascii moved to lib/ui/ and tsdown.config.ts updated accordingly.
No behavior changes; all 58 vitest cases pass and build is green.
