+++ lib/flags/parse.ts
@@ -1,5 +1,5 @@
-import type { Flags, Scope } from "./types";
-import { SCOPES, WORKTREE_SCOPES } from "./types";
+import type { Flags, Scope } from "../types";
+import { SCOPES, WORKTREE_SCOPES } from "../types";
 
 // argv のどこにあってもフラグより優先される
 export type Subcommand = "help" | "logo" | "version";
@@ -23,47 +23,6 @@ export function defaultFlags(): Flags {
     filesChanged: [],
     untouched: false,
   };
-}
-
-export function helpText(): string {
-  return `git-harvest cleans up worktrees and branches based on commit lifecycle stage.
-
-Stages (risky -> safe):
-  files-changed  ->  committed  ->  merged
-
-  A worktree/branch is classified by its most at-risk stage (uncommitted changes win).
-  A flag lowers the threshold and deletes that stage and everything safer; merged is the safe default.
-  "untouched" (no work, identical to base) and "detached" (no branch) sit off this ladder:
-  kept by default, removed by --untouched / --detached (or --yolo).
-
-Usage: git-harvest [options]
-       git-harvest logo
-
-Options:
-  -h, --help                  Show this help
-  -v, --version               Show version
-  -n, --dry-run               Show what would be deleted without deleting
-
-  --committed[=<scope>]       Delete from committed (committed + merged). scope: worktree,
-                              claude-worktree, branch (default: all).
-  --files-changed[=<scope>]   Delete from files-changed (uncommitted included). scope: worktree,
-                              claude-worktree (default: all worktree scopes).
-                              Multiple scopes: comma-separated or repeat the flag.
-  --untouched                 Delete untouched worktrees (no work, identical to base; off-ladder).
-  --detached                  Delete detached worktrees (no branch; off-ladder).
-                              WARNING: a detached worktree's commits are unreachable -- removal can
-                              lose them permanently (no reflog recovery).
-
-  --yolo                      Preset: --files-changed --committed --untouched --detached (all scopes).
-                              WARNING: removes uncommitted changes and detached commits (see --detached).
-
-Subcommands:
-  logo                        Show the git-harvest logo
-
-Invariants are always protected (no flag or --yolo can override):
-  main/default worktree, worktree on the base branch, current cwd worktree, locked worktree,
-  worktree with a running agent session, current HEAD branch, branch checked out in a surviving worktree.
-`;
 }
 
 // argv を Flags へ変換する。各フィールドは args から 1 回ずつ求めるだけで、書き換えはしない
