+++ lib/ui/logo.ascii
