+++ lib/ui/relpath.ts
@@ -1,0 +1,19 @@
+import { homedir } from "node:os";
+
+export function relpath(p: string): string {
+  const home = homedir();
+
+  if (!home) {
+    return p;
+  }
+
+  if (p === home) {
+    return "~";
+  }
+
+  if (p.startsWith(`${home}/`)) {
+    return `~${p.slice(home.length)}`;
+  }
+
+  return p;
+}
