+++ tsdown.config.ts
@@ -2,13 +2,13 @@ import { copyFile } from "node:fs/promises";
 import path from "node:path";
 import { defineConfig } from "tsdown";
 
-// logo.ascii はソースと同じ lib/ に置き、build 時に dist/ へコピーする。
-// 実行時は new URL("./logo.ascii", import.meta.url) で dev / build どちらの隣も指す
+// logo.ascii は lib/ui/ に置き、build 時に dist/ へコピーする。
+// dev (tsx) は lib/ui/logo.ascii、bundle 後の dist/cli.mjs は隣の dist/logo.ascii を読む
 export default defineConfig({
   entry: ["lib/cli.ts"],
   hooks: {
     "build:done": async (context) => {
-      await copyFile("lib/logo.ascii", path.join(context.options.outDir, "logo.ascii"));
+      await copyFile("lib/ui/logo.ascii", path.join(context.options.outDir, "logo.ascii"));
     },
   },
   outputOptions: { banner: "#!/usr/bin/env node" },
